backend/requirements.txt
africastalking==2.0.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==25.3.0
http_ece==1.2.1
idna==3.11
kombu==5.6.2
multidict==6.7.1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-vapid==1.9.4
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pywebpush==2.3.0
PyYAML==6.0.3
redis==7.4.0
reportlab==4.4.10
requests==2.32.5
responses==0.26.0
schema==0.7.8
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.23.0
